feat!: Make log crate optional
- Add log feature flag (enabled by default)
- Move log specific code to `log` module
- Move top level re-exports of log types to `log` module

BREAKING CHANGE: The log crate is now an optional dependency, enabled by
default. The `log::Level` and `log::LevelFilter` types are now
re-exported from the `log` module rather than the crate root. If you
were using these types directly, you will need to update your imports.

```diff
-use clap_verbosity_flag::{Level, LevelFilter};
+use clap_verbosity_flag::log::{Level, LevelFilter};
```
