+++ Cargo.toml
@@ -113,8 +113,12 @@ pre-release-replacements = [
 [badges]
 codecov = { repository = "clap-rs/clap-verbosity-flag" }
 
+[features]
+default = ["log"]
+log = ["dep:log"]
+
 [dependencies]
-log = "0.4.1"
+log = { version = "0.4.1", optional = true }
 clap = { version = "4.0.0", default-features = false, features = ["std", "derive"] }
 
 [dev-dependencies]
@@ -126,3 +130,15 @@ tracing-log = "0.2"
 
 [lints]
 workspace = true
+
+[[example]]
+name = "log"
+required-features = ["log"]
+
+[[example]]
+name = "log_level"
+required-features = ["log"]
+
+[[example]]
+name = "tracing"
+required-features = ["log"]
